% Auto-generated: do not edit by hand
\name{dashSlice}

\alias{dashSlice}

\title{DashSlice component}

\description{

}

\usage{
dashSlice(icon=NULL, iconsize=NULL, id=NULL, loading_state=NULL,
n_clicks=NULL)
}

\arguments{
\item{icon}{Character. The ID used to identify this component in Dash callbacks.}

\item{iconsize}{Character. The ID used to identify this component in Dash callbacks.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional): determines if the component is loading or not
  - prop_name (character; optional): holds which property is loading
  - component_name (character; optional): holds the name of the component that is loading. Assigned by Dash}

\item{n_clicks}{Numeric. Height of the content in CSS Size. This prop is used to center the content between top and bottom of the slice. For example, 2em.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

